[roll] Roll fuchsia [wlan][mlme] Add MLME FIDL -> Fullmac FIDL conversion functions
Implements functions that converts MLME FIDL to the corresponding
WlanFullmacImpl request type, with unit tests.

Unlike existing Banjo conversion code, the new conversion functions
do not truncate the length of vectors or arrays that are too long.
Instead, they return an Error that should be handled by the caller. This
avoids undefined and likely broken behavior if e.g. an SSID or RSNE list
is truncated during conversion.

This is not currently used in production, but will be used when
Banjo is removed from the Fullmac MLME.

Test: fx test wlan-fullmac-mlme
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1078737
Original-Revision: 62aa3f6b283e6a5c08ac16af290d41ea80b9abf4
GitOrigin-RevId: 521a00748e2b9b0d589a5358bccbd778d8dd145f
Change-Id: Ie95b71f84fb60f701cdf2c03499fcaa3057e1e13