[roll] Roll fuchsia [netstack3/conntrack] Plumb TCP information
This CL contains all of the changes necessary to get information about
TCP packets into a new (stubbed) conntrack::tcp module. The idea is that
the core changes for TCP state tracking should be contained to the
conntrack code itself.

The first big change is to replace the TransportPacket trait with a
concrete enum containing the equivalent information for each protocol.
For now, there are only two options, but can be extended to pass
explicit ICMP message information, for example.

Next, there are conversions into SegmentHeader for types used by
filtering. This lets us avoid bubbling up a trait bound on Payload, and
keeps all the data concrete, since conntrack needs to read all of it in
the normal case anyway.

In order to make these conversions possible, we updated packet-formats
and packets to provide getters for a few fields that weren't accessible
before.

Inside conntrack.rs, it's just refactoring to make things cleaner. The
PacketMetadata type is used to hold a tuple and associated transport
data for a packet so it's easier to pass around. We also started using
IP-agnostic TransportProtocol enum in Tuple, since it's much easier to
work with, especially for match expressions.

As part of this, converted a few TCP tests inside filtering and
conntrack to use UDP instead, since the tests weren't relying on
anything TCP-specific, and dealing with TCP is becoming much more
complicated.

Original-Bug: 328064736, 341128580
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1082913
Original-Revision: 9700936a3f0f0a02563ae1071e8029054ceb7b7d
GitOrigin-RevId: 2c1eaf74869a2734a9f15b719fcd86fdcf249c95
Change-Id: I13d10ea8746b0e6a0b7ecdd88b60a2b07a51be6b