[roll] Roll fuchsia [display][banjo] Rename DisplayControllerImpl to DisplayEngine.
Previously the DisplayControllerImpl protocol contained all
Coordinator-triggered methods calls for the display engine.

The "Impl" suffix is not a precise name that describe what
the protocol does -- it's not an implementation of the coordinator
protocol but it controls the engine using a new set of calls.
The "Controller" name is also replaced by "Engine" in new protocols.

This change replaces the "DisplayControllerImpl" to "DisplayEngine",
which is also used by the fuchsia.hardware.display.engine FIDL library.
The corresponding ddk protodef definition and bind constant
are also updated.

Test: no-op. rename only.
Test: fx build
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1087890
Original-Revision: ff39b30db3ffc8effdd738c9114583abb4f515f6
GitOrigin-RevId: ccad6731a8834a89b62eea6d2fdc7741f45ac5b7
Change-Id: I75f67cc47026bd9113d78db50b180508b98ae8ad