[roll] Roll fuchsia [ui] ActivityManager uses LeaseHolder
This change introduces a new struct `LeaseHolder` to handle wake lease
creation, holding, and destruction. For now this is a skeleton class
that does not take a real wake lease from the power framework. It will
be integrated with the power framework in a separate change after
fxrev.dev/1083132 is merged.

This change also parameterizes existing unit tests on the boolean
`suspend_enabled` and adds expectations around whether a "wake lease" is
being held in sync with activity state changes.

Original-Bug: 343256679
Test: fx test activity-manager-test
Test: fx test input_pipeline_lib_tests --test-filter=*activity* -o
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1087514
Original-Revision: f2390001060480d7ed4e6578f90ec713dbbac2d6
GitOrigin-RevId: 8e37e9086cd589c63abf248b567d2a88b0f04122
Change-Id: I77f055f3493ba7c89d6adda287b5f4254c5403f0