[roll] Roll fuchsia [network/test] De-flake DNS timeout test
The flaking test case exercises the DNS resolver's 5s query timeout, and
the behavior that it will return records of one type (e.g. AAAA) even if
it doesn't receive any of another type it requested (e.g. A). The way
it's written is timing-dependent because it advances a fake clock for
the DNS resolver by 1 second every 10 milliseconds of test runtime. This
means in rare cases, even the record that *is* returned by the fake name
server doesn't arrive and get processed in time before the 5s timeout.
Increase the ratio of test runtime to fake DNS resolver time so the
response is more likely to be processed.

Original-Fixed: 355083806
Multiply: netstack-dns-integration-test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1089512
Original-Revision: 82cdc02b5178a6736bac030b85c9a3f309126787
GitOrigin-RevId: dea6748c3761bf2ffd4272d47a0f3c76a5b50ca5
Change-Id: Id32abe871792859b87a6588d67e23813074582c3