[roll] Roll fuchsia [Lacewing] Move Suspend-Resume test from stable to unstable
Suspend-resume based tests started flaking (because of platform issues)
since fxr/1051364. Since then test has not met the updated stability
criteria (which is 1000+ pass runs). Even though this test is in stable,
it was not running in infra as along with fxr/1051364, this builder was
also demoted from CQ to FYI as part of tqr/843553. So correct steps to
take now include:
1. Move all suspend-resume tests from stable to unstable (this CL)
2. In another CL, update the builder back from FYI to CQ so that other tests that are part of this builder can be run in CQ again

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1089912
Original-Revision: 878638acea0560d7fa79731bd3f52d86d3a6e802
GitOrigin-RevId: 1add127e32c160ac4f4b9f33e1cb074a08368752
Change-Id: I2917b0f183cc04f6c0750162592a9b8bdd57f2cc