[roll] Roll fuchsia [fidlc] Add better error for name not found with versioning
This improves the fidlc error message when a name is not found in a
versioned library. Before, it would just say "name not found". Now, if
the name actually is found but in the wrong version range, it lists all
the version ranges and locations it did find the name.

Test: fx test fidlc-test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1077906
Original-Revision: aa034e2add340406bc01376d8cafdbf5df99e5ec
GitOrigin-RevId: 116a9fdae25026882ece219e7ccb5b91aedd5e78
Change-Id: Ia1655fd6555943bcdb0853694411f007c9edcb2c