[roll] Roll fuchsia [reland][starnix][vfs] Prevent umount when files are open
If there's a file open inside a mount, we are supposed to prevent the
mount from being removed by returning EBUSY from umount. To count the
number of clients, we add a marker to the Mount object that we reference
from each ActiveNamespaceNode. It's important to hold this reference
from ActiveNamespaceNode objects rather than FileObjects because the cwd
and root fields on the FsContext need to keep the mount from being
removed and those fields are NamespaceNode objects rather than
FileObjects.

Test: umount02, mount_test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1089740
Original-Revision: 22023551f9cd0335e5e86f0d580a37ae50687218
GitOrigin-RevId: 37ceb927addf0b9c52b40b35b6c33de5938f169b
Change-Id: Idf2f040e9d61326ed9f1e77638844616918f2436