[roll] Roll fuchsia [ffx][pkg] Support for managing server process instances
This adds a struct to manage instance files that contain
the information used to start server processes.

This information is used to implement stop and list commands
and could be used to identify the appropriate package server to
use for a command based on the project and the target device.

Original-Bug: b/342205668
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1051464
Original-Revision: dc4ba45cdce96bed1daaab4ee13d6502466a97ad
GitOrigin-RevId: da562bec2cffb5ed3785fc732b85560f93754626
Change-Id: I5d9ab0d62e916a01ff5e6353d64ecf5ddedecf8b