[roll] Roll fuchsia [sys/test_manager] integrate debug_agent
This CL integrates debug_agent into the test environment to enable
the generation of backtrace files in the test output when tests fail
or time out. For the time being, this feature is disabled due to issues
with exception channel conflicts. Once those conflicts are resolved,
another CL will enable this feature.

When enabled, backtraces go to a stderr file for the suite, which is
shown symbolized in the report that goes to the console when running
'fx test'. Using the stderr file instead of a special file allows the
backtrace to be displayed without host-side changes. A future CL will
introduce a new artifact type for this purpose.

This CL causes Rust backtraces to appear twice in the fx test console
output, because Rust backtraces are delivered using a different
mechanism. This issue will be addressed separately.

Original-Bug: 286555768 303314982

Test: unit tests added

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1038975
Original-Revision: b6efb25b5fc9e900802d06fb3c4ea06a2af251cb
GitOrigin-RevId: 35f2e31c53559fa0ccd41be365ce15a59cb124d2
Change-Id: I025e71440ad81754a971e9b43a1f923d6f870a82