[roll] Roll fuchsia [ffx] Fixes a bug with Fastboot mDNS discovery.
The fastboot mDNS discovery code would only check a single record to
determine if the domain is correct, rather than all records. This would
lead to flakes where the record we were interested in would not be
detected.

This also adds mdns_discovery to the list of tested libraries in ffx.
For some reason it was not included so likely wasn't being tested in CQ.

Original-Bug: b/353345798
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1089926
Original-Revision: 0977c485a234b4dbd2a3dce1e105c4045b360a51
GitOrigin-RevId: 748e7829ee2bf642827aae357fd48cd0684b8fd1
Change-Id: I2eb68486231736b87f8c69a7fcacf285207222ab