[roll] Roll fuchsia [pkvm] Ship pkvm boards in the minimal builders
The pkvm-x64 and pkvm-arm64 boards are now built and uploaded to CIPD in
continuous integration.

In order to ensure unique board names, we have to override the `name`.
The board_configuration.gni had to be modified to use the overridden
name instead of the target_name.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1090806
Original-Revision: 13d4a0b2ba7e8aa6980da60e732004bbaf113ef3
GitOrigin-RevId: 4139965290c98c0290aa1cec68db35030b9340d6
Change-Id: Id31ef81e6a20796e21428c53d110177bc653b967