[roll] Roll fuchsia [broker] Preserve dependencies during transition
On transitions between power levels, elements may be in either state and
the broker should attempt to defer decisions to activate or deactivate
claims until the state is well-known. The broker now checks to see if a
claim is satisfied only if it is satisfied at both it's current level
and it's transition level. It also does not drop claims that are
currently transiting.

Further, this change does not send required level updates until the
element replies to a previous required level change. This ensures that
only one transition is ongoing for an element at any given time.

Original-Bug: 342450114
Test: fx test src/power/broker
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1065215
Original-Revision: 0e68dffa3608b7c43d4560b6107fb7b88a631c9a
GitOrigin-RevId: e006a7f507813267b6b3b4c3ba424816d112dc08
Change-Id: I61c53499361ef6da139310bc528ecdd47d339886