[roll] Roll fuchsia [kernel][phys] Ensure the correct allocations will survive into hand-off
There are two cases around BootZbi::Load() in which we were ending up
with duplicate allocations with type kKernel and kDataZbi:

(1) In physboot, we Load() a kernel ZBI alone, yielding a spurious data
ZBI allocation from its empty tail, which we then overwrite.
(2) TrampolineBoot allocates staging kernel and data ZBI images for the
purposes of trampolining, but these clash with the fixed-address
versions we reserve ahead of time.

We expediently address both of these situations with simple freeing or
recharacterization of types post-Load(). This is easy enough to do,
readable, and in (1) involves codepaths that are not long for this
world.

As logging feedback that the correct kKernel and kDataZbi allocations do
end up surviving, we add these types to those included in the desired
set of hand-off allocations.

Original-Bug: 347766366
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1081985
Original-Revision: 56f0dafbbd62a16ad4a5f87c195fe1fe27850557
GitOrigin-RevId: 90b69d9133a12998ef7bd18c15dd347374a8efa6
Change-Id: I0eda5621c5b8e8cb7bb3d0c347a7aed6bd9b8ddb