[roll] Roll fuchsia [dwmac] Fix race with Ethernet Start
In an attempt to remove netsvc from all images, we observed vim3 had
failures because the driver never propagates the online status up to the
legacy Ethernet driver if the child driver ends up binding after the
link is observed up by dwmac.

Fix that by making sure to notfy the child as soon as it binds, which is
a known workaround on the legacy Ethernet interface.

Original-Bug: 355280494
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1091712
Original-Revision: 6d9b7205b565f0188656b36ab2c0ca8b00e04a59
GitOrigin-RevId: ef8f83b77995e9950d7576be28a30891785d9308
Change-Id: Ib61f73853293f5b3e738daa326361938f4b1e5f1