[roll] Roll fuchsia [kernel][x86] Use memmove for copying kaslr-fixups.
The code that emits the relocation fixups for kaslr is copied out from
the bss section during `_start`, into the location that's right after
the bss section.

If the kernel contains too many relocations, as is the case for some
kernel variants like `kubsan`, the relocations will partially overlap
the destination buffer. This caused an issue with the previous code for
copying out the relocations: by copying from left to right, it would
overwrite the same code it is copying.

The solution is to copy right-to-left if the destination buffer overlaps
with the source buffer.

Original-Bug: 328825932
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1068314
Original-Revision: 9a15c09bd2fb202d760d67041b2503480e115cf2
GitOrigin-RevId: b563526542736e0d584f1fd56bae20a6b8de8ccb
Change-Id: Ibe21b5dcaf261df1213fc0b5aa6445ed41b6352c