[roll] Roll fuchsia [routing] remove capability_provider from FilteredAggregate
`CapabilitySource::FilteredAggregate` was the only remaining
(non-phantom) usage of generics in `CapabilitySource`. This was due to
the `capability_provider` field, a `Box<dyn>` of a trait defined in the
routing library.

There were only two different values ever placed in this field:
`OfferFilteredServiceProvider` and `OfferAggregateServiceProvider`. This
commit removes `CapabilitySource::FilteredAggregate` and replaces it
with two different variants. `CapabilitySource::FilteredProvider` holds
the information necessary to construct an
`OfferFilteredServiceProvider`, and
`CapabilitySource::FilteredAggregateProvider` holds the information
necessary to construct an `OfferAggregateServiceProvider`. With this
change, there are no remaining (non-phantom) instances of generics in
`CaapbilitySource`.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1090595
Original-Revision: dbdf0594088aff145c457477f0c35b86f579fc62
GitOrigin-RevId: c3716446af7c70ec3a5bcd64731a3c9251b2ce80
Change-Id: I19ecfe57f23f104f7cacbedfaaf64b03e1ad867b