[roll] Roll fuchsia [starnix] Propagate EAGAIN from std::thread::Builder::build
Currently starnix panics if std::thread::Builder::build fails, but it
can fail with EAGAIN when thread creation fails due to insufficient
resources. Propagate this particular error variant to userspace.

This requires careful handling to ensure that the TaskBuilder provided
to the spawned thread is always explicitly released; we introduce a
one-shot channel to provide the builder to the thread only in the case
that the thread was successfully spawned.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1086597
Original-Revision: 4e4908d7b0c4369088d2e1fcc6ba4f67c68c47d7
GitOrigin-RevId: 40602a24531fa4866374285eae98c6502f17456f
Change-Id: Ie929e25e19e6edf6575798ae85ae353a0fce7f59