[roll] Roll fuchsia [swd-e2e-tests] Try to first flash with product bundle
This teaches the OTA tests how to download a product bundle, and then
attempts to use that product bundle to flash a device. Unfortunately we
still need to support flashing builds as old as F11 though, which don't
support PBv2, so we will fall back to the flash manifest if we can't get
the product bundle.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1090792
Original-Revision: 1297d9eca47764e40f9ef73d7dae5c140307de26
GitOrigin-RevId: 1f74e3e5e934f881488166410926488f1b92119e
Change-Id: I44109997a6bb5cfd072edf0fa491a5c175cd67eb