[roll] Roll fuchsia [graphics] Fix type of PFN_vkInitializeOpenInNamespaceCallbackAddr
This function was returning pointers for no real reason, so fix the
type. The return value was completely unused, so this won't affect
anything except UBSAN; UBSAN is disabled in this patch until the
follow-up that fixes it in the Vulkan-Loader repo.

Original-Bug: 356893475
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1092874
Original-Revision: adf244d7f09b81517b5dd00a680865f9aa7701fd
GitOrigin-RevId: ef153eaf89c09cf7d127b2d2248a80a0b2529ee1
Change-Id: I8b0d391be6dae6ffca0f6e18f9da18767c15c518