[roll] Roll fuchsia [starnix] give view_id annotation with PresentView
This allows window managers to identify the view using the annotation if
they so choose. Note: We eventually want to get the annotation value from a config to correspond with fuchsia.wm.InitialViewIdAnnotation. That change will be a follow-up to this one.

Original-Bug: 349124304
Test: manual test shows that boot still presents the frame as expected
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1091412
Original-Revision: 05274166a4945f2917fb10855b2773b2ffdf1a1f
GitOrigin-RevId: a5bb2234a263fe772f7e9bac89062d11e17e2e46
Change-Id: I5f5a56e11daca07cce45d14e4570db78df15579d