[roll] Roll fuchsia [kernel] Convert IdlePowerThread to boot time
Converts the idle power thread to use boot time when transitioning the
system into suspend. This required changing some tests as well.

Note: This change does not pause the monotonic clock on suspension, so
the boot and monotonic timelines will never diverge. Thus, this change
should be a functional no-op. It is being made solely to minimize the
amount of code that must be checked in behind a feature flag.

Original-Bug: 341785588
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1094013
Original-Revision: 9db77ec34b90fd07c5d933ddcde736d0afab39a5
GitOrigin-RevId: 878bb2ee0266eb82eaaeadf0ddfb25b53b8bb7b3
Change-Id: Iebccd52ed3e3c6c223ffb054a25f0f44c2e5b24f