[roll] Roll fuchsia [kernel][vm] Page request status is first page
Makes the signalled status of a page request be the status of the first
page, instead of any failed portion. This allows for the case where the
first pages in the request are supplied successfully and later pages
are failed to be correctly handled.

Run-All-Tests: True
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1091097
Original-Revision: bdc4c69550c9051a0dabc16ee83dcba17be74807
GitOrigin-RevId: 0bd36b697c49acb559e4de70cdf98d59c8df3d99
Change-Id: I104fa7d9059626a03d56680a0a00c81d65a2f688