[roll] Roll fuchsia [netstack3] Lookup Multicast Routes during IP RX
When handling a received IP Multicast packet, query the routing table
and determine which targets the packet should be forwarded to, if any.
For conformance with Linux/NS2, attempt to forward multicast packets
regardless of whether the host has membership in the multicast group.

For now, the packet isn't actually forwarded.

Original-Bug: 353329136
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1092872
Original-Revision: 65438002325ace6367980cdd0e597cb95b8cc887
GitOrigin-RevId: 0f7616d15dfe4d534f45d1084a819e4ed15e684d
Change-Id: I1be56ba59ab73411f8d66b4e224f92a6f1f3c249