[roll] Roll fuchsia [bazel-sdk] check for extra resources in package checks
The package_checker_tests were only looking for expected manifets and
did not check ot see if manifests that should not be included were being
included. This allowed some of the generated cml files to be included in
the package under the wrong name. This was not breaking the packages
since the file was also included under the correct name but it caused
confusion and was just not correct.

BUG: b/356932372
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1094235
Original-Revision: eaee2d065ca059ce37358d861552307270f2001a
GitOrigin-RevId: 860a5cb543e95e19de2b0afb8b89621e6aae095a
Change-Id: Ia924d43e841a40a5a5d60d8588bc7c1f8752a3d2