[roll] Roll fuchsia [bt-host] fix test_component names in build files
The bazel fuchsia_unittest_package is bundling the component manifests
twice. One of the manifests is named after the component and one is
named after the generated file even though they are both the same source
file. We are cleaning up this logic but we need to first update the
usages of the incorrect name.

BUG: b/356932372
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1094236
Original-Revision: ac24f60884ad6d8b6c166c33ad8f6ea0742f4f9a
GitOrigin-RevId: fa88b9dde446bfdcfa134f00b5a171c4b2a4f2d3
Change-Id: I0dbd41f0cf6a592f5ee26e31fc91a29c71b9222f