[roll] Roll fuchsia [audio] audio_driver_tests validates VMO rights
The audio driver test suite now validates that the VMO
handles returned from an audio driver has the required
rights. In order to do this, the test must reason about
whether the ring buffer is 'outgoing' or 'incoming'.
This is trivial for Dai and StreamConfig drivers (the
driver declares the `is_input` bool in its properties)
but for Composite drivers it must understand the ring
buffer element's place in the signalprocessing topology.

Test: audio_driver_tests -- --admin

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1093754
Original-Revision: f0d3326a452cc81cba9d7ca019800f07741794fc
GitOrigin-RevId: 907ceaf8ac9c97506dac4c0e21cd895e091db91e
Change-Id: I0cf41b79fbb82afb1bc41f5f796d264325b5af19