[roll] Roll fuchsia [kernel][phys] Introduce formal end of hand-off
This change introduces a formal end to the hand-off phase via the
function `EndHandoff()`. It is currently expected to be called in
userboot, unsets gPhysHandoff, returns the remaining hand-off data left
to be consumed (e.g., the ZBI in the physmap), and populates its subset
of userboot's VMO handles.

Previously, the handle population logic was built into the
instrumentation VMO preparation code, as that seemed a convenient spot
to put it at the time and there were similarities. This is now naturally
moved out to a hand-off-specific site.

Moreover, as of this change, <phys/handoff.h> is now evaluated within
the userboot toolchain.

Original-Bug: 357155617
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1094015
Original-Revision: ce355c1606745412832b99356dbfc1bc7e8bd976
GitOrigin-RevId: 2b486e117af45e7ab636975b9a8994d66972e0db
Change-Id: I206cfc97b5c2edc02b80db9e54ffe93fd39b1406