[roll] Roll fuchsia [routing] define CapabilitySource in FIDL
This commit duplicates the rust CapabilitySource enum (and its
associated types) with an identical structure in FIDL. The traits
cm_rust::{FidlIntoNative,NativeIntoFidl} are defined on
CapabilitySource.

This allows the CapabilitySource type to be easily serialized and
deserialized using fidl::{persist, unpersist}, and thus allows us to
delete the lengthy amount of code that was manually converting this type
into a sandbox::Capability and back.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1091934
Original-Revision: 2cd4163df52196bf1a75d88253629a49aa7ff2c3
GitOrigin-RevId: 6c9d7517b96d28d422d887f23927e5b2be4bcb5b
Change-Id: Iee7311cfa31c9b7698b621391d2fe8e85e4a709a