[roll] Roll infra/recipes [lucicfg] Ignore unknown fields when validating configs
A recent lucicfg roll (http://tqr/870660) broke the lucicfg validator
because it starts setting a new commit-queue.cfg field that hasn't yet
been propagated to the version of the CQ config proto file that's
vendored in the recipe engine repo.

This field (and most new potential new fields) are not relevant for the
purposes of this validation, so we can safely ignore them so that
lucicfg rolls do not get blocked when they set new fields.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/1095652
Original-Revision: 0401d9b14524f5102cb681d27607917695b013fb
GitOrigin-RevId: 1d38803b5daa379a79b3bed95dd7a299db7b93b9
Change-Id: I5d906e03378890cc68e9356a34c64438685a14b0