[roll] Roll fuchsia [display][coordinator] Always bind the coordinator ServerEnd to Client.
Previously in some coordinator tests, the coordinator server-end channel
was not bound to a `Client` instance but kept as a standalone ServerEnd
and stored in the DisplayControllerBindingState type.

This change removes the test-only constructors for Client and
ClientProxy that store the ServerEnd,  and replaces them with
`BindForTesting() / InitForTesting()` methods that bind the ServerEnd
to `Client` instead. This unifies the client binding logic and makes it
possible for us to remove the `DisplayControllerBindingState` type used
to emit events.

Test: no production code affected.
Test: display-coordinator-unittest
Original-Bug: 42075581
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1081902
Original-Revision: 541ef430f096fe3b8230c6f392320a192f47622a
GitOrigin-RevId: 177afbe483d9a84f71c18532fd299cee013f21a0
Change-Id: I7f9f728a94f67c294ce31ad49095e0b2d518528b