[roll] Roll fuchsia [fxfs][testing] Add poison pill option for Fake device.
Adds a poison pill option for FakeDevice teardown. This allows us to
coerce a stack trace when the last reference to a device is released
rather than crashing in the thread that is checking for the other
references to have been released.

Original-Bug: b/356752755
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1095814
Original-Revision: c4e5bb1ea259b8ce49e176cd9081d7873fbb995b
GitOrigin-RevId: 54bd4138a04cb6dc468fc018ff6f4ccfb536b184
Change-Id: I42358187e52a828d7e6d01f11c9833dfd5108d07