[roll] Roll fuchsia [bazel-sdk] decouple autogenerated cc_tests
This CL removes the autogenerated cc_tests from the build graph when
they are not needed. We were always placing the generated components in
the build graph because they were added to the FuchsiaUnitTestComponentInfo
provider. This made it so the aspects that collect resources would pick
up the manifests incorrectly.

Additionally, we updated the fuchsia_component_manifest logic to use the
component_name for the output name if provided which matches the in-tree
logic.

BUG: b/356932372
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1094237
Original-Revision: 589cc1bb8ab476c2d6ef93f01de1551dd2fc03eb
GitOrigin-RevId: 11ba45019bacf7b0f031a9ed061561a42e5c62bc
Change-Id: I3f877449e1d8b65de9bc6425d65924176552e856