[roll] Roll fuchsia [windowed-stats] Implement serialization
Serialize all the information needed to deserialize the data from a
TimeMatrix. Adjust the allowable types for a TimeMatrix to match
what's in the design document, as otherwise we don't have a type
descriptor to include in the serialized buffer.

Update the `interpolate` method to no longer return a serialized buffer.
This way, we provide an option to update the TimeMatrix without having
to incur the cost of serialization.

Original-Bug: 350797382
Test: Added unit test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1084873
Original-Revision: 34b18f41206bffff528ba49dcf5e6f1b8a518b86
GitOrigin-RevId: 5e79c1192384b563d26a1fdbc7b3a659d735ce28
Change-Id: If768d82694ae0cd2825f45321cb4293e989bc2ba