[roll] Roll fuchsia [ffx] Load external subtools on demand
Instead of scanning external subtools only to invoke a built-in
plugin, wait on the I/O until we actually need to scan them.
This may help with a situation in Infra in which there is a
delay when the daemon starts.

Original-Bug: 316626057
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1095741
Original-Revision: 36b10be511dfa9b371543101b018bd5a25318818
GitOrigin-RevId: d84870be94dbef990f33895a2dee87943867ab05
Change-Id: I264c0b1657b295c859135bacd843c7a5ce870909