[roll] Roll fuchsia [media] Remove duplicate routes
Routing LogSink manually is not needed,
as this is done implicitly through the client shard.
Future CLs will update the routing used in the shard,
so it's necessary to cleanup these duplicate routes
in order to update the shards.

Original-Bug: b/345827642
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1096899
Original-Revision: 23dc217c20ca541641051fb2aa419b580b7dc848
GitOrigin-RevId: aef68267ecb8ed9e2e481c030b74e251f097e998
Change-Id: I75c63734bb163b133d1e4d2bc88668c68d87a0f0