[roll] Roll fuchsia [fshost] Remove duplicate routes
Routing LogSink manually is not needed,
as this is done implicitly through the client shard.
Future CLs will update the routing used in the shard,
so it's necessary to cleanup these duplicate routes
in order to update the shards.

Original-Bug: b/345827642
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1096897
Original-Revision: 98d7ab97132c34711189074f9560f9a2eaf556ea
GitOrigin-RevId: 90a81c80aae6319b7560c9109ebbdaaf8df750dd
Change-Id: I3a552abe879627aba6b76cda82c158b7506b940a