[roll] Roll fuchsia [ui] Remove duplicate LogSink routes
Routing LogSink manually is not needed,
as this is done implicitly through the client shard.
Future CLs will update the routing used in the shard,
so it's necessary to cleanup these duplicate routes
in order to update the shards.

Original-Bug: b/345827642
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1096893
Original-Revision: 3c16e4e3e7a485160d107649043e154946a63643
GitOrigin-RevId: b2b4d6e2bfacbd6fbf44153d499c676c1534c694
Change-Id: I5f3d6833eac6ee880a998ef4579e33f5d9cfeff3