[roll] Roll fuchsia [abi-compat] 25% speedup through conversion refactor
This inclues some changes to how we convert from the IR types to the
comparison types. The use of FlyStr is significantly reduced because it
was showing up high in traces and it has properies (like thread safety)
that don't matter here. Instead of allocating and copying Vecs to track
state and nesting as we walk through types, we use references to values
on caller's stacks. These paths are still turned into Vecs with Strings
when storing them on the converted values, and there's more room for
improvements there.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1094064
Original-Revision: 376827683a8f2d5f8be86c6e1e34c7fe658b6ff4
GitOrigin-RevId: 4db0775cd3969c8d597c2c7dc59562947ba93e98
Change-Id: If4790e9af89ef4c0c8a3b39f85ed54f363bdcf7a