[roll] Roll fuchsia [sestarnix] Update init_child_process security state
Previously tasks that were injected as children of init were using the
kernel's security context. This was causing tests to fail a bunch of
security checks. In production tasks are not created this way.

Now, we instead create the security context with `task_alloc` on the
init task.

A future change will allow test components to specify this more
directly.

Original-Bug: 355809979
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1097576
Original-Revision: 25edc5af8cd38bf4482639731b101077e3395c4f
GitOrigin-RevId: df616b81b3f938c6c9488cbde05a56e9974a28c6
Change-Id: I257965e5c174d8511b4a370ef75ebebd5f08c1a7