[roll] Roll fuchsia [netstack3/conntrack/tcp] Closing
Like with Established, this is a thin wrapper around
do_established_update.

There are two things we're not planning on handling (for now).

The first is that we aren't going to handle data segments after a FIN is
sent. See the comment at the top of Closing::update for more details on
this. Since it's not handled by gVisor or Linux, it doesn't seem
important for us to, and shouldn't affect correctness.

Second, we are not implementing a TimeWait state. That state is not
required for correctness, as is further described by the comment at the
end of Closing::update where the transition would occur.

Original-Bug: 328064736
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1091353
Original-Revision: a383ccd6f6a5a89adda2ce603fddff18353494d9
GitOrigin-RevId: 289071c228f3e8566d73a8ed8f60e3e30ba77990
Change-Id: Icfdbb40762a1f2e350c16b9a05f81d6143d58279