[roll] Roll fuchsia [netstack3] Fix dangling neighbor timer
Currently, when a neighbor entry is created in the INCOMPLETE state, the
timer to retransmit a multicast neighbor solicitation is unconditionally
scheduled, and only then is the packet to be sent out queued. However,
serialization of the pending packet is fallible, and if it fails, the
timer that was scheduled remains in the timer heap; if a future timer is
scheduled for this neighbor, it trips an assertion that expects no other
timers to be scheduled.

Fix this by only scheduling the timer once serialization of the pending
packet was successful.

Original-Fixed: 357115127
Test: netstack3-core-ip-test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1097992
Original-Revision: 552f9409230bfe5877b6e612c4e295e10f8f6837
GitOrigin-RevId: 6765a6705faa48c88ec4fdd7d8f908e63de23275
Change-Id: Ibb3efa394807bb0103c0ebe599b651e486cbe156