[roll] Roll fuchsia [starnix][selinux] Revise logic for FsNode security label choice
fs_node_effective_sid() always returns a SID, suitable for use
to access-check the file. It uses the SID attached to the FsNode
if any, otherwise delegating to fs_node_resolve_security() to
attach one.

fs_node_resolve_security() chooses a SID based on the file-system
labeling scheme and mount options, including attempting to resolve
labels stored in extended attributes, if appropriate. This process
is fallible but if a SID is resolved then it is attached to the
FsNode and returned to the caller.

Original-Bug: b/355587237, b/334094811, b/349800754
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1095872
Original-Revision: 8c1048c65b188537e00aea1a0aaee90dd9bd2119
GitOrigin-RevId: cb082af984600d441780422629c77af3974d21d5
Change-Id: I545d9ea57afaac98be498a1a06a288397add7713