[roll] Roll fuchsia [vfs][cpp] Bind SynchronousVfs connections under lock
Similar to ManagedVfs, ensure we call Bind() while still holding the VFS
lock. Not holding the lock was leading to flakes on debug builders when
checking that the connection itself was still owned by the container,
since there was a small window where the VFS could be shutdown before
Bind() was called.

This also ensures that the BindImpl() functions in each connection type
do not dereference nullptr when accessing the dispatcher associated with
the VFS.

Original-Fixed: 358451182
Test: fx test fs-connection-test
Multiply: fs-connection-test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1070953
Original-Revision: c6455a536f4445dabbdf62edfea5b861fb06bdd6
GitOrigin-RevId: 995928b29b65bcb0685f3ea1322e348249caaec9
Change-Id: If70c7ea4836fa648f3081c60e6089e5a345e2786