[roll] Roll fuchsia [driver][test-realm] realm-builder remove shard
This removes the process launcher offer to the driver test realm as it
is not used after the dfv2 migration. This also allows us to remove the
shard that we had which only had this entry in it.

Since the shard itself is accessed in vendor repos we will keep it in
this patch, then when those are removed we can delete the shard.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1098075
Original-Revision: 3570231330d7637fd32abd2db17158a560e30c5d
GitOrigin-RevId: bf0fa2dea0c40df2263cf0b36e8c26c3b459768f
Change-Id: I8c893ea98b642bd5f360901266e0e3726052b137