[roll] Roll fuchsia [driver_test_realm] Use realm_builder_exposed_dir to forward connections
Prior to this change, we were forwarding capabilities individually, but
if we use a directory, we can do them all in one go.  This simplifies
the code with a minor increase in verbosity in the manifest.

This fixes an issue with exposing the fuchsia.hardware.ramdisk service
which wasn't using the forwarding mechanism, and also allows us to
use dtr_exposes rather than the deprecated exposes Start argument.

Original-Bug: b/357779890
Original-Bug: b/358475423
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1098232
Original-Revision: 11072292270914eb9afc68b66851bd9d4f5ca9d6
GitOrigin-RevId: 6001523034cdc5f6c2432b37e63d1a57825ecd3a
Change-Id: Ia80e7752e5a80999c871ed0674a7d730b558b400