[roll] Roll fuchsia [netstack3] Introduce rules table in IP state
This change introduces the lock levels and accessor traits for the
top level rules table. Currently we don't install rules from the
Bindings, so the routing decision is still made from the only
default rule in the table which is just using the main table. There
is no change in behavior yet, we rely on the existing tests to make
sure.

Original-Bug: 354724171

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1095732
Original-Revision: 03922769c63753301e0f9465b7b6fc3a72c43d48
GitOrigin-RevId: 31218f112c1583fa5aa0fd06d47e3dd26eedf5f8
Change-Id: I2f445df68a003aae4b3e561bd6342d0a8de9a2cc