[roll] Roll fuchsia [usb-hid] Use left-shift operator to calculate power of two.
The driver was using the pow() function to calculate the polling
rate. It invoked libc math functions which are restricted in
drivers.

This CL changes the driver to use the << operator to calculate
the polling rate so that it won't call the libc math function
anymore.

Original-Bug: 359565197
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1100372
Original-Revision: 3ef025774d824d6db01c675f194c0bee488aba0f
GitOrigin-RevId: f109e23c02bd32c0f6d0bdd7076fdc435b2f61e3
Change-Id: I985daf74554611ee2bdf4f9e161e035a5ab36feb