[roll] Roll fuchsia [kernel][pi] Use latched version of `now`
During each of the PI related operations, a current "now" time (on the
monotonic timeline) is latched at the start of the operation and
passed to HandlePiInteractionCommon to be used as the definition of
"now" for the rest of the operation.

However, during ThreadBaseProfileChanged and
UpstreamThreadBaseProfileChanged, the value latched and passed to the
common handler was not the value being used by the callbacks used to
maintain the dynamic scheduling parameters.  Instead, a new value of
`new` was being computed during the callback using
Scheduler::CurrentTime().

Fix this by capturing the value of `now` computed at the start of the
operation in the lambda currently being used for the callbacks for the
base profile changed operations.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1093079
Original-Revision: be0c02c5e0565f323cc426de6fe95db06d22c8d0
GitOrigin-RevId: 2e5256e72c2b5f75cfe625315300179cfa42dd5f
Change-Id: I57ba7e5990e9560e744580e65a3413ccab4bd2ca