[roll] Roll fuchsia [wlan][telemetry] Log WLAN connectivity time series
Log time series for:
- wlan_connectivity_states
- connected_networks
- disconnected_networks
- disconnect_sources

Create a `metadata` node to include the state names for
`wlan_connectivity_states`. The metadata for `connected_networks` and
`disconnect_sources` are also moved into the `metadata` node.

Also update the `connect_events` Inspect log to refer to the
`connected_networks` metadata.

Every telemetry interval, connectivity state is logged to ensure no gap
in data and all time series are interpolated so that telemetry doesn't
have to do too many interpolations when the time series are requested by
Inspect. Note that the telemetry interval introduced here is 10s,
which is different from the 15s seen in wlancfg. So when we migrate
wlancfg to use this library in the future, we will need to reconcile
this difference (https://fxbug.dev/356238939).

Original-Bug: 331389728
Test:
1. Updated unit tests
2. Verified Inspect data locally

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1088952
Original-Revision: 2df147c87bf734d57a6867ece506610c41c6875b
GitOrigin-RevId: 3ca0b154dc50ee0f5133f9c3db486c85d2170277
Change-Id: I52f85a915c39af84de48a9762540210bcfeeb4df