[roll] Roll fuchsia [bt][gap] Integrate LegacyPairingState with PairingStateManager
This CL wires up the legacy pairing process code to the rest of the
system via the PairingStateManager.

Depending on the pairing events we encounter before/after the ACL
connection is complete, or depending on the result of interrogation with
the peer, we will tell the PSM whether to use SSP or legacy pairing.

Original-Bug: b/42173830
Test: fx test //src/connectivity/bluetooth
Manual Test:
- Paired and streamed audio to/from Nelson DUT with iOS and veho devices
  supporting SSP to ensure SSP still functions
- Paired and streamed audio to/from Nelson DUT with iOS and veho
  devices. Cherry-picked fxr/1078858 to simulate legacy pairing

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1062293
Original-Revision: 31a9d97a99e1571ad674961a0647b3de85d9265e
GitOrigin-RevId: fdf6faa7543964875c685cdf754ecfc2e3ed588c
Change-Id: I4e1d940ee4486c8aef940a63afd9fe5ca2c4cd9a