[roll] Roll fuchsia [sysmem][assembly] Move sysmem driver to bootstrap
This change allows the microfuchsia product to avoid including the
sysmem driver. It's possible that sysmem should be moved to it's own AIB
and boards/products should opt into a feature that causes sysmem to be
included, but for now, all products other than microfuchsia need it, so
that can be done at a later point.

Original-Fixed: 335548998
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1101919
Original-Revision: 54a64bca0622eb036d88e0c727e584b615a81b76
GitOrigin-RevId: a2ff98ed3b12a034c49d706a45bf3ea4c606957a
Change-Id: I56974a58e4a03ea3fe044e8e9ce6790c68e51b04