[roll] Roll fuchsia [fdomain] Simplify Async error handling
We previously had separate messages for "an error occurred" and "reading
stopped". This was to support transient errors during which reading
might continue.

In the prototype we found no place to emit such errors. Furthermore,
having every error simply stop reading and require it to be restarted
seems like a minor issue, so now we have only one "stopped" message
which may or may not have been caused by an error, in which case it will
indicate so.

Original-Fixed: b/356729179
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1097373
Original-Revision: 9d78bb5a2781900e957dff5ec4198f42a088516d
GitOrigin-RevId: d949e28b5f1a1b32d3a8b40c77294fcbb7f49acb
Change-Id: I71897027e943eb489c295de52fe095c66a0ce1fa