[roll] Roll fuchsia [ffx][fho] Expose generalized daemonize
This exposes starting a command as a daemon process.

This is added to support long running tools such as repository server
to run in the background in the same way the ffx daemon is run in the
background. This includes the libc call to make the daemon process, and
also set up the logging to capture the output of the daemon process.

Original-Bug: b/358390369
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1101152
Original-Revision: 4f077633a7087853a307f4942b99c580198fe1b7
GitOrigin-RevId: c5265d2b1ed790395d7ca61994a6f562fb53727d
Change-Id: I001f24f922d4b5d8e64f79e6818a977a40510bfe