[roll] Roll fuchsia [ctf] Reland "[ctf] Add component_manager tests to subbuild"
This CL fixes the underlying problem with the test. Including the
component_manager with a config for running in a nested setting also
caused a "data" directory to appear, which is asserted in the test. This
change now adds a placeholder config which results in the data directory
appearing, allowing the existing assertion to pass.

This reverts commit 9093dc6b2d61a538e5c631688317027b704a0bb0.

Run-All-Tests: true

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1103312
Original-Revision: 4deee296fc738e0f9e0f26027ad2b1648cf7971f
GitOrigin-RevId: 3869295641e62e875cb39829afe10084af98b5bf
Change-Id: Ic163a0bb0a4dc12937d17719ebe5b52a4b1847e7