[roll] Roll fuchsia [kernel][vm] Improve CSM lock annotations
Changes the ContentSizeManager::Operation to have a constant parent_
references and use a separate valid_ flag for tracking validity. This
allows for the external users of the operation to used AliasedLock
acquisitions instead of having to assert that the lock is held.

Asserting that the lock is held is still needed for when the CSM itself
is iterating over the operation queues.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1103752
Original-Revision: 4315f1fa9857c331a64e4a51e930425726bc6b88
GitOrigin-RevId: 787c1395d1e1e3d81990bc2ab2752e0758beac3a
Change-Id: I0b6b1e98d12015261d154cc34770a248d9cf4126