[roll] Roll fuchsia [dtr] provide add function for dtr_exposes/offers
This provides a function to add the exposes and offers that are being
provided through the dtr_exposes/offers fields. Also adds the feature to
the c++ variant of the realm_builder.

This change also improves the consistency between the routes we add with
the various realm builders and the driver test realm component manifest.

Original-Bug: 358475423
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1098077
Original-Revision: c70145931e206e488e7ca8f5b06376ce19ecde35
GitOrigin-RevId: cecb8b33aa3253861db4516a1cd5f1e1397cbe2d
Change-Id: Ifb378c17c632c0a34697a6e4c27be3b1b3c72ad5