[roll] Roll fuchsia [bazel] add --dir support to refresh-bazel-compdb
This CL adds a flag which lets a user pass in a directory that will be
searched for targets that can be added to the compilation database. The
directory will query for all valid fuchsia packages and run the update
on those targets.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1104555
Original-Revision: 7120cb61ab93ca0921aeabe31831a8e7aa340d39
GitOrigin-RevId: 07049a43f9b3dbff8c750e0c5ff2579195f0e5cc
Change-Id: I2080ac495c540292f2294ac8830b28f2dda07423