[roll] Roll fuchsia [ffx][repository] Add IfThisThenThat lint for devhost literals
For the usages of the default `devhost` repository in ffx, add
IfThisThenThat linting to help developers in case they want to change
this default in the future.
Note that there is a current limitation that the linter can't have
multiple targets in the ThenChange(path) clause.

Original-Bug: b/359534719
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1103656
Original-Revision: 45a44e667898a2dcf1e90a7c2528fd20ce2b6495
GitOrigin-RevId: 81483f5f056a27528ed084f9186c1892879b6455
Change-Id: I17878f47f3bb5252cd24fa7397d831e4887fa13d