[roll] Roll fuchsia [starnix][remotefs] Report symlink target size as st_size
Ensure we query and set the content size as well as timestamps for
remote filesystems that deal with symlinks.  Update fxfs to ensure we
report the link target size as the content size attribute so it will be
reported as st_size.

Original-Bug: 356676802
Test: fx test starnix-tests --test-filter="*symlink*"
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1097114
Original-Revision: 6750ad75c9e75920c8562db1cb7433fc603a5c8d
GitOrigin-RevId: 017d831ca0cb7e056a169a315247a85e86bf03dd
Change-Id: I0696c443e47c38390f9c638bb67055dfae43527d