[roll] Roll fuchsia [netstack3] Ignore multicast bound addr for send
When a socket's bound address is multicast, don't try to use it as the
src address for sent packets. Instead, follow traditional source address
selection.

Note, this new functionality exposed an existing problem in the netsvc
ping tests: Echo Requests sent from the netstack to a multicast address
are looped back to the sender by default. With this change, the Netstack
is able to respond to its own Echo Requests, which interferes with the
expectation of observing a reply from netsvc. The fix is to disable
MULTICAST_LOOP on the ping socket.

Original-Bug: 360187268

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1102472
Original-Revision: 756ff2558aea31ff2e3baab8b8823cbca2c2d912
GitOrigin-RevId: c59da3b383966494da77b21dc2abb4c6407bd150
Change-Id: Icee34f64813c71950cccb14baa91e26f5aa85fb5