[roll] Roll fuchsia [ffx] Move SSH error checks to frontend. Report to user.
This takes ssh errors encountered when attempting to get a remote
control proxy and sends them to the user. Previously these errors would
be logged to the daemon log and then removed, not giving clients the
insight into what may be hanging.

Now, the error is always reported to the user, with reattempts happening
the same as before, but just on the client side. If a deemed-non-fatal
error is encountered, at which point a reconnect attempt will happen,
the error is printed for the user to see.

There are three considered-fatal SSH errors: key verification failure,
permission denied, and invalid arguments. These are all errors that
cannot be fixed no matter how many reattempts occur.

Original-Bug: 357958365
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1101154
Original-Revision: 1f1fd5228ba3a2f719fd83b743c0764a2f57b2b7
GitOrigin-RevId: ce0905d1005ba14f5464c7cb4c809d936d8329c2
Change-Id: I1ea7bde2744dc3942dfeed0151a606b695ddd87e